ui: Fix presubmit
The Protobuf typescript library we use has an option for using
long.js to handle numbers that exceed the range representable
via JS. We don't use this and never have however we generated
Typescript bindings which allowed for long.js to be used. This
forced us to coerce to the number type in many places and caused
many problems with tslint. This sets an option on the binding
generator to remove the long.js support letting us get rid of all
the casts.

Change-Id: I7ea85dcc40b260797fafa666133fe2b3b989623c