tp: fix slice tracking behaviour when multiple S event occur b2b
Catapult had some unfortuante behaiour when it came multiple
S slices were seen back to back; it simply overwrote the slice
information for the (upid, cookie, name) tuple.

This allowed for some very badly behaved trace points to go
undetected in Android which take advantage of S events by emitting
multiple back to back (without finishing any of them) and expecting only
one of them to be tracked.

To deal with these sort of clients, introduce a legacy unnestable code
path into slice tracker which can track these sort of slices by only
allowing a single slice for a given tuple to be open at a time. We also
track how many times we see this behaviour as an arg.

Bug: 160969198
Change-Id: If4f2fc8f480b61510f5668492c3341a0bf2fbedd