Port the queuing delay computation over to TraceProcessor.
Second to last metric to port over. This one is used in various places
since we've focused on queuing delay. This allows the metric name be the
same between metrics, and we can then port over the blocking task
computation to a query that will work inside scroll_jank_cause using
this metric as a base.

Change-Id: Iebfef4d335ad67395a1bb416d641f794ef8bf0a0