ui: fix counter track quantization
Currently, counter track quantization is very misleading as we only
draw the value with the max(value) for each range which means we miss
sharp drops as well as obscure whether a counter was at more than
one value. As well as this, we don't take into account last(value) in
a quantum correctly.

This CL changes the frontend code to draw the whole min-max
range (as a line) when more than one value is collapsed to a single
pixel. This essentially matches the behaviour of the old (prod) code but
in a more efficient way.

We don't use the last_value function as it can only be used as a window
function which introduces a 2x regression in the query. Instead, we
define an aggregate equivalent of it which does basically the same thing
for our purposes.

Bug: 161248499
Change-Id: Id0626c8e0bc17b41e31fd2264847818b73a93ba1