ChromeConfig: Add a setting for exporting trace data as JSON
Add a setting to enable legacy JSON exporting in Chrome. This will be
used by Chrome's Perfetto backend to perform the JSON conversion in the
(sandboxed) tracing service instead of requiring client API consumers to
perform the conversion by themselves. This setting will only be
supported by the client API backend in Chrome.

Bug: 158460807
Change-Id: I069d3df5570d222415f948dcf08bf74102526aa0