Add new metric for queuing delay scroll jank cause.
This metric finds slices that could have blocked the next step in the critical
flow of a scroll from executing. It does this by looking at the period of time
between steps collecting all slices and figuring out the overlap. This allows us
to count how often a certain task appears between janky and non-janky scrolls.

In the process of porting our current version of this metric and adding the diff
test I found 2 bugs related to including the current slice stack which means we
were inflating numbers for events that happen as a part of the critical flow.

Bug: 159606819
Change-Id: I0cf7d01b542e36dca49987b80729bd407877f4a7