ui: Reduce special handling of the summary track
As part of revamping how tracks work it would be nice to fix a couple
of long standing issues:
- Track groups can't be nested
- Pinned tracks lose a lot of context (e.g. which process they're from)
- Can't reorder tracks within a track group
- Can't add/remove tracks from a group or re-sort the groups by another
  scheme.

At the moment the summary track is treated specially by a lot of the
code which leads to a lot of complication both within the code and
conceptually as we address the above issues, for example:
- If I move every track out of a track group what happens to the group?
  It seems like it should disappear but then what happens to the summary
  track?
- Since we must choose a summary track we sometimes choose one that
  isn't useful (e.g. we show the scheduling summary track when there are
  no scheduling events).
- If make a new group and drag some tracks into it how should I decide
  the summary track?

A simple scheme for handling this is: whichever is the top track in a group,
that is the summary track. This solves or avoids many of the above issues:
- Since there's no special summary track we can disappear empty groups
- At least something useful will appear in the summary slot
- Simple rule for which track is the summary track which is (hopefully)
  obvious and controllable to the user after little experimentation.

As a little first step along this path this reduces the special space for
summary track id in the state.

Change-Id: Idb3358e348a642f4f0e4be82ba183fa120ab7973