Tests: Add client library + service stress test
Add initial version of the client library & service stress test.
This tests tries to push the producer & service to its limits
by spawning a configurable number of processes and threads,
each writing events at a customizable rate.

Some work is still left, I still want to tune the configs and
reason on what to expect. Meanwhile checking in the harness.

Change-Id: Iea0b01e52018125e44f42c5b61c687c60315893b