Handle multiple batches when converting to pandas.
Currently we weren't handling multiple batches when it came to parsing query results into a pandas dataframe. This commit fixes that plus tests have been added for the as_pandas() function to ensure correct behaviour. This patchset also raises TraceProcessorException if an invalid cell_type is sent for processing (tests also added for this).

Change-Id: I3fe791fb9f054ad572e11ab001d98925213c9274