tp: rework thread state in prep for parsing blocked reason
Because blocked reason appears *after* waking but *before* wakeup, our
approach of adding rows straight away simply does not scale.

Instead of this, store the descheduling and wakup information until the
next sched event to add this to the table. This is safe because we
always sort by timestamp at the end so we don't strictly need to insert
in order. Also, make sure to flush at the end to cover all the pending
events at the end of the trace.

Bug: 164449656
Change-Id: Id7d18a2a9beb267deb00459395a743dfbf7d72cc