perfetto: fix documentation for unix socket and simplify ipv6 handling
This CL fixes the documentation of unix socket which incorrectly stated
that Listen always returns a socket instance. Unfortunately, this was
not true and caused subtle crashes when code relied on it. Instead
document that bind and socket creation failures result in nullptr being
returned.

Make trace processor http handle this case and also simplify a bit of
the ipv6 handling which turned out to be unncessary.

Change-Id: I4a8653d2b2285e6535802c3a5108ac3ef7609bd8