Ensure that our join key doesn't contain "NoJank".
We were joining on metric_name but metric_name includes if its janky or
not. This meant that janky rows always had a null on their LEFT join and
were replaced with '0' (we assumed they didn't have a corresponding
non-janky duration). This fixes that as well as modifies the test to
ensure that both janky and non-janky rows have the same values for one
particular event I computed the average for manually.

Bug: 162577602
Change-Id: Ie06ec1bf6242297885d01fc483e6f7833e8e6b91