ui: Use id for thread state selection
Originally thread state slices were derived and so didn't have a concept
of a slice id. Now we have a thread state table with ids in the TP so
we can make the selection consistent with other selections.
This also makes thread states work correctly since the quantization
broke some assumptions selection was relying on.

Bug: 168035025
Change-Id: Ic5aca7fb62eebf593c6895bf24136ded319c75f3