ui: Fix build dependencies
We don't keep accurate track of the outputs of tsc compilation. This
used to be fine: we manually put the index files and since tsc
compilation re-wrote every file this was enough to trigger the
dependees to rebuild. However once incremental compilation was
enabled (for speed) only *some* of the files get rewritten. Which meant
eveything worked great if you touched an index file but not if you
modified anything else.

Fix this by adding a stamp file to the outputs which is always written
no matter what.

Bug: 120010518
Change-Id: I3245a36f6a95ca6067e2772154b58c9996bb3322