gn: make generate_descriptor consistent between chrome and standalone
Previously, Chrome used to use generate_descriptor relative to to the
proto_out_path while Perfetto used to be relative to the target
directory. This led to inconsistency in where the descriptor was being
generated.

Use the perfetto_proto_library macro to make this consistent between the
two by always specifiying the full path of the descriptor relative to
proto_out_path

Change-Id: Ic7761194bccbe524cdf50986266ea6098ce52a0d