Require core layout for time in state metric
Computing the time in state metric when the CPU core to cluster mapping
is not present results in incorrect values.

Introducing another level of aggregation to handle the case where all
cores are labelled as "unknown" adds complexity. Instead return no data
for the metric, because data is not very useful without the core type
breakdown.

Bug: 168725744
Change-Id: Ica5e224eb3a8799ac9b5d8139cd338555c026479