Two small fixes for pages with a 16KB system page
These two tests are failing on 16KB systems because the test code
makes wrong assumptions. Specifically:

BufferedFrameDeserializerTest.FillCapacity:
It initializes BufferedFrameDeserializer with max_capacity = 16k,
but then this CHECK that fails:
PERFETTO_CHECK(max_capacity > base::GetSysPageSize());
Which is fine on 4KB system pages but fails precisely for one byt
on this new platform with 16KB pages.
Solution: make that > a >= :)

PagedMemory::Allocate:
creates a 4KB PagedMemory and expects that a memory access @ [4KB+1]
gets trapped by the mmap(PROT_NONE) red-zone.
The issue is that the red zones can only be page-aligned, so on a
16KB system 4KB+1 is a valid memory pointer because falls into the
0th 16KB page.
Fix: in the test replace kSize = 4096 with GetSysPageSize();

Change-Id: Ia19fe932b7d8fdc5a908551c15438a45efd33a96
Bug: crbug.com/1128455
Test: I don't have access to 16KB-page hardware, but I debugged in my
      mind and it works there.