perfetto-ui: Fix details panel height
When switching to a different page and back to timeline, the details
panel will begin at default height.
I decided it was probably not worth storing the previous height.
Also simplified the logic. And ensured a tab is always marked as active
even if the currentTab doesn't match any tabs.

Bug:159523702
Change-Id: I550ad0ad30578a9f7a186eabc5ef0bf6a098f017