tp: wait to fill buffers when parsing gzip before sending to next stage
This CL changes the gzip parser to wait to fill the full 32MB chunk
before sending it onto the next reader. This prevents massively
overallocating memory when small chunks are passed in.

Also simplify DecompressTrace by making full use of the class instead of
trying to just use a static method.

Bug: 169885963
Change-Id: I6d7144e23d47aed6fa2c8bef21bbef5df06cf604