tracing: Don't try to observe events before connecting
In TracingMuxerImpl, don't try to observe events from the tracing
service before the connection has completed. Otherwise the observation
request will be silently dropped on the floor.

The old code happens to work with the in-process tracing service because
the consumer connects immediately, but with a system producer there is a
longer delay.

We also fix a faulty DCHECK() that tries to dereference a null message
field.

Bug: 170126760
Change-Id: Ie8b2eb35a2c57a2f091b071b30d241c81120ec6b