tracing: Add debug check for duplicated interned data indices
If an interned data index is defined in an anonymous namespace
that is included from multiple translation units or otherwise
duplicated between files, we can end up with multiple copies of
it in the same program. Because they will all share a memory
address through TLS, this can lead to subtle data corruption if
all the copies aren't exactly identical. Try to detect this by
making sure the Add() function address remains constant.

Bug: 156727531
Change-Id: I4c22781d4c22ea97e96fd1eed6e4b8656bbac32d