UI: allow to re-share postMessage traces echoing back the opener
We cannot re-share traces opened via postMessage() because doing
so could create a trace that outlives permissions of the original
trace.
Instead, when the opener pushes a url together with the ArrayBuffer,
echo back the URL. If the opener implements it right, the passed
URL should cause an idempotent window.open, having a similar effect
to the permalink (% not recovering our UI state).

Bug: https://github.com/google/perfetto/issues/14
Screenshot: https://screenshot.googleplex.com/AvSVghLuXMx9aLX
Test: https://jsfiddle.net/nahck81j/1/
Change-Id: I918de9761ef5d32f4f7abb4b2ed5f24ef74168d6