ui: Show pages for system probes when chromeOS selected
Added isCrOSTarget() method which returns true when chromeOS
is selected from dropdown in GenConfig() added checks for
isCrOSTarget and in that case if true-this method will return
object which contain all (selected) system probes in dataSource.
In Record menu added check for isCrOSTarget and if true on
recording page all system probes will be shown.

Bug: chromium:1141391
Change-Id: Ibd5c16a8ed07de6309dd800cda65ea19a7806e6e