[flow] Replace flow_eof field with terminating_flow_ids field.
Instead of writing a bitmask of which flow_ids should be terminated,
separate ones which should be terminated into a separate
terminating_flow_ids field.

This allows us to get rid of 64 flow ids per track event limitation, as
well as be more efficient for a common case of a pair of out / in flow
track events by avoiding the extra field in "terminating in" case.

R=eseckler@google.com
Change-Id: Ieeb3de192e36db602fc16f8241b6a66c1760f6b7