Handle erroneous cases in Frame lifecycle phases
As per the base R branch, acquire fence is tracked only for buffers that
are successfully latched. This behavior has been updated in QPR to track
acquire fence for all the buffers that were queued, irrespective of the
latch status. In cases where the app is overperforming, the next buffer
is queued before surfaceflinger tries to latch. This causes an
incomplete phase in the UI since the events sent are Dequeue1, Queue1,
Dequeue2, Queue2, Acquire2, Latch2. This results in an incomplete phase
slice for buffer 1 since we start the slice at Queue and end it at
Acquire.

To fix this issue, this change closes any unclosed GPU or App slices if
a Dequeue event is seen. Since we do not know how long the slice is
actually supposed to be, the duration is made to be -1 so that the UI
can ignore this slice.

Bug: 171828838
Test: unit tests
Change-Id: I822923f1267b6eb92026e90ce5efa08d73658a15