traced: provide options to set socket permissions.
This change adds options "producer-socket-group",
"producer-socket-mode", "consumer-socket-group" and
"consumer-socket-mode" to set group ownership and file mode bits of the
producer and consumer sockets in traced.

Example: ./traced --producer-socket-group traced-producer
--producer-socket-mode 0660
starts traced and sets group ownership of the producer socket to the
group "traced-producer" and sets file mode bits to 660 (rw-rw----).

Bug: 147789115
Change-Id: Ic15828116aebcbd213174bdb75163f4a0841d775