Query results pane: single-click vs double-click
Change the default behaviour of clicking on a row of the query results
table from selecting a slice and switching to 'current_selection' tab to
selecting a slice and leaving on the current tab (useful when you are
jumping around search results).

The old behaviour (select slice and switch to a new tab) now requires a
double-click.

Also use |id| instead of |slice_id| to look up the slice (as
|slice_id| is non-documented column and internal_slice, ancestor_slice and
descendant_slice do not have it).

Also update |frontendLocalState.currentTab| when selecting the first
tab.

Change-Id: I688efcfb4436fb6ca8018b0b3f4b85fd936de285