PRESUBMIT.py: update arg name to avoid warnings on newer depot_tools
Newer depot_tools have switched the argument naming convention again,
and cause presubmits to print warnings even when using
allow_list/block_list:
  presubmit_support.py:744: UserWarning: Use files_to_check in FilterSourceFile

The logic is still backwards compatible (the allow_list is used if the
new param is empty), but the warning is undesirable, as it is pure
noise.

Afaiu, as-is, this patch will force people to upgrade to a depot_tools
that understands the new parameter name. Which feels like an annoyance
for such a small change. One alternative would be to drop the argname
qualification, and use purely positional args for now (then it'll work
with either depot_tools version).

Change-Id: I2302928ffd7863b2429e5f33e55ffc07134d7101