Do not wrap unwindstack::FrameData.
This way, we do not need construct a wrapped object.
We also do not need to destruct the vector returned by
libunwindstack in DoUnwind.

We want to change libunwindstack to take an existing
std::vector<unwindstack::FrameData>* to re-use. This is
much easier if we don't wrap in our own type.

Bug: 169938761
Change-Id: Ia127f0e62944b9042d7a4ac01c52807a22b12389