UI: fix 'Cannot read property 'source' of undefined'
The bug was about sidebar.ts hardcoding state.engines[0].
The engine ID keeps changing monotonically as more traces
are open. The proper way to deal with that is to do
Object.values(engines)[0] as we do elsewhere.
In future we should get rid of this as we never open more
than one trace at the same time.
This CL also fixes a bug in the WASM error reporting code.
This was causing some bugs happening when an exception is
thrown in a ccall, to be reported as '{}' (e.g. b/168645606)
rather than showing the full stack trace. Turns out that
JSON.stringify doesn't do anything useful for Error().

Test: manual: open two traces, share the second.
Bug: 173785472
Bug: 173016918
Bug: 172517820
Bug: 172517541
Change-Id: I0b3db9fc51cbc2adca64a01f5610dd4254b99928