ui: Select correct column when first value is null
In the columner format if one (or more) of the initial
values in a column are null all of stringValues, longValues, and doubleValues
are partially filled (with '[NULL]', 0, 0 respectively). From the point
we see the first non-null value only the relevant column is filled.

Previous to this CL this could lead us to select the wrong column to
iterate over causing bad errors for our callers since we broke
guarantees we promised the typesystem.

Change-Id: I4937cfc6fcf2faebe4b3e7699674a19cda7f892d