tracing: Handle disconnections in TracingMuxerImpl::SyncProducersForTesting
TracingMuxerImpl::SyncProducersForTesting is called between tests to
make sure all the producer data sources have been registered with the
new tracing service for the current test iteration. Because the producer
might not have yet noticed the previous system service having been
terminated, sending the Sync here can be what triggers disconnection at
the IPC layer. The producer will automatically reconnect in response,
but notification of the disconnection arrives asynchronously, we can't
immediately tell if the sync succeeded or not.

To work around this, we sync all producers twice. The first round is
needed for detecting disconnects, while the second round ensures that
reconnection has completed.

# While compiling Chrome in the background.
Test: perfetto_integrationtests --gtest_filter=PerfettoApiTest/PerfettoApiTest.TrackEventCustomTrackAndTimestamp/System --gtest_repeat=1000 --gtest_break_on_failure

Change-Id: Id10940942f39b70c45a5d7a7de2caa040f361587