TraceProcessor: add streaming query support.
This CL adds an endpoint to issue queries and retrieve
results in row-oriented batches. The batched data is
designed to allow fast parsing from JS/TS by packing
all cells of the same type together.
This CL changes the /query HTP endpoint to return a
Transfer-Encoding: Chunked sequence of batches. The binary
format is still unchanged.
This endpoint is in reality temporary because the whole TP
interface might be switched to websockets soon anyways to
allow pipelining and proper chunking support.
For now, however, we need a way to get this data in the
UI to make progress there.

Bug: 159142289
Change-Id: I584fb94f8326d82144ca5c55b2da060b56637348