Add new dynamic table to look up ancestors of CPU stack samples.
This removes the need for a recursive query which is a significant
performance bottle neck.

I validated the results on a test trace with results here:
https://docs.google.com/spreadsheets/d/1whwjzHbU7QQ1iaR3W-dqLRRBxl5LAY5NKnBeRmidFCk/edit?usp=sharing&resourcekey=0-5u4EtMsfmDyD1X_lvLi9CA

The new method just skips a row in the validation which was repeated
in the recursive query (I think this was a bug? please double check
though).

Trying on the example trace in the bug and selecting a stack sample
from ChromeIOThread over the IOHandler::OnIOCompleted the performance
was similar. This does not bode well for this fix.

Bug: 168056830
Change-Id: I89f2e042d0d4dc43f424272e71aced4c0c426346