metrics: Add test for cpu_by_combined_rail_mode table
Adds a simple python generator that creates two processes with main
threads and emits RAILMode slices as well as cpu counters on each to
check that the table is properly populated.

Due to b/175110985, this also increases the trace times for the
combined_rail_modes.py test by a factor of 1000.

Bug: 174764142
Change-Id: I00350879456138f0b503a13d4f9c38e210bdeb61