Introduce PlatformHandle
In UNIX-based systems any system resource we deal with is an
"int" (a file descriptor). On Windows the situation is more
complex because there are (At least) two types of resources:
1. HANDLE objects, coming from long-standing winapi.
2. int resources, coming from the more modern posix-compat
   layer (e.g. _open, _close, etc).
This CL introduces PlatformHandle to deal with types that
are still int on UNIX but a HANDLE on Windows.


Bug: 174454879
Change-Id: Ief3ea5d6a79952c1f1577cd0515476818bbf3d92