Add SysUI CUJ jank causes to Perfetto UI
See https://screenshot.googleplex.com/3ptf23RHFFzdMRq

Because frames are drawn on multiple threads there is some overlap in
their timings. Perfetto UI does not seem to support showing slices that
are overlapping in a single track, so we are aligning all root causes
with the main thread Choreographer#doFrame slice which is not ideal, but
ensures the root cause slices never overlap.

Bug: 161781469
Change-Id: Ib5f60e77e7113210059bb89b834c3132a16d2a6f