metrics: Add test for real_power_by_rail_mode table
Adds a simple python generator that creates two processes with main
threads and emits RAILMode slices as well as power rail data to check
that the table is properly populated.

Also fixes incorrect usage of SPAN_JOIN due to overlapping slices for
each power rail by partitioning over the power rail. Due to b/175287144,
this is done by first converting the power rail name to a unique integer
index, as non-integer indices are not yet supported.

Bug: 174764142
Change-Id: I28f82c3456c1a92e7a4b73b679ae6c24d43310e5