metrics: Use touch events in modified_rail_slices
The "animation" RAIL mode is the default bucket and is used even during
user-driven scrolling when that scrolling is handled directly by the
compositor thread.

This changes the metric to look at InputLatency touch events and
whenever there is an active touch slice (or it is in between "begin" and
"end" slices of certain touch event sequences like GestureFlingStart and
GestureFlingCancel) intersecting with an "animation" slice, it creates a
"response" slice.

Since the result of this could create adjacent response slices, this
also adds in a new post-processing phase at the end that merges adjacent
slices with the same mode.

Bug: 172558170
Change-Id: Ibeacc5056b726b1f7d6e9f3368815844337d5e87