base::Pipe: Use HANDLE on Windows
This CL switches base::Pipe from using the POSIX-compat API _pipe()
to CreatePipe(). Turns out that is more proper for use in
base::Subprocess (which will be introduced by an upcoming CL) because
CreatePipe() returns objects that can be waited for using
WaitForMultipleObjects, and hence by the TaskRunner.
This CL also:
- Fixes ScopedPlatformHandle on Windows. Turns out both nullptr
  and 0xffff... are invalid values.
- Introduces some file_utils methods to Read/Write from
  HANDLEs.

Test: ran perfetto_unittests on Win and Linux.
Bug: 174454879
Change-Id: Ic8b51821a5909857ddb437b02de3e642594453af