Win port: add base::getopt_compat
Windows doesn't have <getopt.h> but all our executable these
days depend on it. We have two roads ahead:
1. Create an abstraction layer, like base::ArgParsers to deal
   with cmdline parsing.
2. Create a getopt emulation for Windows.

1 is too much risk and too much effort. On Android our cmdline
is de-facto an API, we can't risk subtle breakages.
Instead it's easier and lower risk to create a getopt emulation
for Windows.
This CL also introduces a test that checks that the behavior of
the two is consistent.

Bug: 174454879
Test: perfetto_unittests --gtest_filter=GetOpt*

Change-Id: I68752e615ad0aaf24226ddbce0d7ba77140bdca9