Fix the chrome thread enum to string conversion
The chrome thread types are not ordered in the proto definition and the
array of string IDs stored were always out of bounds for some threads.
Also fixes the ordering of FG_BLOCKING and BG_BLOCKING type names that
were displayed incorrectly.
Adds network service thread type to the list.

Change-Id: I6f3de1a18865def11c3f40debf9f4dec91ed4999