Make preceding_flow and following_flow follow relative slices
Based on altimin@'s cl:
https://android-review.googlesource.com/c/platform/external/perfetto/+/1484245

This changes the preceding_flows() SQL function to follow flows incoming
flows by checking all the ancestors of the slices that are followed. The
following_flows() function will follow all the descendants of the slices
followed. Removes connected_flow function because it can be trivially
implemented in SQL. The connected_flows() function is renamed to
directly_connected_flows() and will follow flows without relatives,
and this is kept as is for the UI to show flows.

Change-Id: I941a7e9bd1bd8328108fa0091906c546e181f9ec