perfetto: add skip probability option for triggers
This adds the ability to specify a probability of skipping a trigger to
the trace config. This is used in the service to probabilistically
decide whether a trigger should be allowed to finalize a trace.

This is useful for reducing the instance of high-frequency triggers
from finalizing a trace at the expense of lower-frequency triggers.

Change-Id: I43a54c15309d0034f090c002c4961520b4b0d3f3