traced_perf: cfg proto: add optional event filter when counting tracepoints
A filter is a stringified expression consisting of several operators and
the tracepoints' fields. E.g. "prev_pid == 0 || next_pid > 42". This is
set via an ioctl on the perf_event_open'd fd, and the validation is done
by the kernel itself. Invalid filters cause the entire config to be
rejected.

Bug: 178961752
Change-Id: Ie30a1008fc69d89b38bbb6389b0b30aeaf65361f