UI build: use custom build script, support --watch
This CL switches the UI build from GN+ninja to a custom
JS-based build script. The main reason for this is the
fact that incremental builds are incredibly slow today
and that is unsolvable with any build system which is
based on invoking tsc + rollup on every change.
This instead leverages the --watch mode of tsc and
rollup.

Change-Id: Ifc87abeed609a89b8bde8349870e43d17ba6ec33