UI: move initialization from index.html to JS
This CL removes all dependencies from index.thml to
version-specific code in the UI. This is to get into
a state where we never touch index.html again, so can
be stable across versions.
As part of this, the CSS loading is moved to the
frontend initialization, together with the CSP
initialization.
Deferring the CSS initialization has unvelied
a "global constructors"-type-of race in
css_constants.ts, which was trying to read back
the computed stylee without waiting for CSS to be
loaded.
This CL also adds a spinner in the index.html, which
will show up for the time it takes to load
frontend_bundle.js, and a BSOD-type error page
for unrecoverable errors that happen in the early
frontend bootstrapping stage.

Bug: 179864115
Change-Id: I8e3dfd7726237f1a6a66ce79e69f81dccdf92ab6