TrackEvent: Make the category index a runtime parameter
This patch changes the track event category index to be a normal
passed-in parameter instead of having a separate template instance of
the trace point machinery for each category. This has the following
benefits:

  1) Reduced binary size. For example tracing_module.o[1] becomes 6%
     smaller on x64 with this change. Each individual trace point,
     however, grows by one instruction to pass the category index
     parameter.

  2) Better cache locality through reusing the same tracing code across
     trace points.

  3) This makes it easier to wrap TrackEvent with a C interface where
     the set of categories isn't known at C++ compile time, i.e., they
     are defined outside Perfetto.

Note that the category index is still computed at build-time like before
for trace points with statically known categories. This patch only
affects how that category is propagated to the track event internals.

Data sources using custom trace point traits need a small modification
to be compatible with this change:

  // Before:
  struct MyTracePointTraits {
    static constexpr std::atomic<uint32_t>* GetActiveInstances() {
      ...
    }
  };

  // After:
  struct MyTracePointTraits {
    struct TracePointData {};
    static constexpr std::atomic<uint32_t>* GetActiveInstances(
        TracePointData) {
      ...
    }
  };

[1] https://source.chromium.org/chromium/chromium/src/+/master:third_party/perfetto/src/tracing/test/tracing_module.cc?q=tracing_module.cc

Change-Id: I71ddd8010cdfec8332d81d621075ac0ff1432648
Bug: 160140065